+++ packages/auto_test/lib/src/generator/url_generator.dart
@@ -2,18 +2,17 @@ import 'dart:math';
 
 class UrlGenerator {
   static const _protocols = ['http', 'https', 'ftp'];
+  final Random _random = Random();
 
   Uri generate() {
-    int index = Random().nextInt(_protocols.length);
+    int index = _random.nextInt(_protocols.length);
     String protocol = _protocols[index];
 
-    bool hasPort = Random().nextBool();
-
+    bool hasPort = _random.nextBool();
 
     return Uri(
-      scheme: protocol,
-      host: 'auto.test',
-      port: hasPort ? Random().nextInt(99999) : null
-    );
+        scheme: protocol,
+        host: 'auto.test',
+        port: hasPort ? _random.nextInt(99999) : null);
   }
 }