+++ packages/auto_test/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.2.1
 
 - feat: support string params
+- feat: support uri params
+- feat: support datetime params
 
 ## 0.1.0